requirements-dev.txt
mypy==1.7.1
pydantic==2.5.2
pydantic-settings==2.1.0
pylint==3.0.2
pytest==7.4.3
pytest-mock==3.12.0
coverage==7.3.2
pandas-stubs>=1.5.3,<2.0.0
types-sqlalchemy==1.4.53
types-python-dateutil==2.8.19.14
